requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.0
Django==2.2.6
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
ipdb==0.12.2
ipython==7.8.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.730
mypy-extensions==0.4.1
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.0
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.6
wcwidth==0.1.7
wrapt==1.11.2
xmltodict==0.12.0
zipp==0.6.0
